Handle mixer responses correctly
Use a function to handle mixer responses on stream action calls correctly.
Before, there was no differencations between responses from api calls
that didn't end well.
Now, we print a different message if the api/verb is not found and if
the request was transfered but an error occured during it.

Change-Id: Ib4f9c3cf14b6612cd14762b74466ce484fdfb09d
Signed-off-by: Jonathan Aillet <jonathan.aillet@iot.bzh>